Implemented better handling for changed packages.json files
Currently only packages with a change in configuration would be deployed in case the config file changed - regardless of whether or not packages with unchanged configurations had changes.
Now packages with changed config will be deployed regardless of what else happened and packages without changes to their config will be undertaken the standard change test.